refactor: Moved KF, Gen3, ICGC into new packages
