+++ score-client/src/main/java/bio/overture/score/client/manifest/icgc/DownloadManifestReader.java
@@ -15,43 +15,45 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN                         
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package bio.overture.score.client.manifest;
+package bio.overture.score.client.manifest.icgc;
 
-import static com.google.common.base.Preconditions.checkState;
-import static java.nio.charset.StandardCharsets.UTF_8;
+import bio.overture.score.client.manifest.DownloadManifest;
+import bio.overture.score.client.manifest.DownloadManifest.ManifestEntry;
+import com.google.common.base.Splitter;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableList.Builder;
+import com.google.common.io.LineProcessor;
+import com.google.common.io.Resources;
+import lombok.NonNull;
+import lombok.SneakyThrows;
+import lombok.val;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.util.List;
 
-import lombok.NonNull;
-import lombok.SneakyThrows;
-import lombok.val;
+import static com.google.common.base.Preconditions.checkState;
+import static java.nio.charset.StandardCharsets.UTF_8;
 
-import bio.overture.score.client.manifest.UploadManifest.ManifestEntry;
+public class DownloadManifestReader {
 
-import com.google.common.base.Splitter;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableList.Builder;
-import com.google.common.io.LineProcessor;
-import com.google.common.io.Resources;
-
-public class UploadManifestReader {
-
+  /**
+   * Constants.
+   */
   protected static final Splitter LINE_PARSER = Splitter.on('\t').trimResults();
 
   @SneakyThrows
-  public UploadManifest readManifest(@NonNull File manifestFile) {
+  public DownloadManifest readManifest(@NonNull File manifestFile) {
     return readManifest(manifestFile.toURI().toURL());
   }
 
   @SneakyThrows
-  public UploadManifest readManifest(@NonNull URL manifestFile) {
+  public DownloadManifest readManifest(@NonNull URL manifestFile) {
     return Resources.readLines(manifestFile, UTF_8, new ManifestLineProcessor());
   }
 
-  private class ManifestLineProcessor implements LineProcessor<UploadManifest> {
+  private class ManifestLineProcessor implements LineProcessor<DownloadManifest> {
 
     int count = 0;
     Builder<ManifestEntry> entries = ImmutableList.<ManifestEntry> builder();
@@ -70,23 +72,31 @@ public class UploadManifestReader {
     }
 
     @Override
-    public UploadManifest getResult() {
-      return new UploadManifest(entries.build());
+    public DownloadManifest getResult() {
+      return new DownloadManifest(entries.build());
     }
   }
 
   protected ManifestEntry createEntry(List<String> values) {
     int i = 0;
     return ManifestEntry.builder()
+        .repoCode(values.get(i++))
+        .fileId(values.get(i++))
         .fileUuid(values.get(i++))
+        .fileFormat(values.get(i++))
         .fileName(values.get(i++))
+        .fileSize(values.get(i++))
         .fileMd5sum(values.get(i++))
+        .indexFileUuid(values.get(i++))
+        .donorId(values.get(i++))
+        .projectId(values.get(i++))
+        .study(values.get(i++))
         .build();
   }
 
   protected List<String> parseLine(String line) {
     val values = LINE_PARSER.splitToList(line);
-    val expectedColumns = 3;
+    val expectedColumns = 11;
     val actualColumns = values.size();
     checkState(expectedColumns == actualColumns, "Expected %s columns found %s for line '%s'",
         expectedColumns, actualColumns, line);
