+++ score-client/src/main/java/bio/overture/score/client/metadata/legacy/LegacyMetadataService.java
@@ -15,8 +15,10 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN                         
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package bio.overture.score.client.metadata;
+package bio.overture.score.client.metadata.legacy;
 
+import bio.overture.score.client.metadata.Entity;
+import bio.overture.score.client.metadata.MetadataService;
 import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
