+++ score-client/src/main/java/bio/overture/score/client/storage/gen3/Gen3Client.java
@@ -1,4 +1,4 @@
-package bio.overture.score.client.storage;
+package bio.overture.score.client.storage.gen3;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
